Normal font size for dataview table headers
Tables created with dataview plugin have headers with larger font size. It is down to personal preference, and main motivation for this change is that I personally do not like it, but also, practically, larger font size is very unusable in dense tables, because it causes cells to not fit properly.